requirements.txt
asgiref==3.5.2
Django==4.0.6
django-types==0.15.0
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
